Script: Lua: Removed compiler warning
git-svn-id: file:///home/michael/svn/yuni/trunk@835 7e728022-c0ab-4603-8e65-c666ead28a33

+++ src/yuni/script/lua/args.cpp
@@ -15,7 +15,7 @@ namespace Bind
 
 	int Argument<Script::Lua*, int>::Get(Script::Lua* ctx, unsigned int order)
 	{
-		return lua_tointeger(ctx->pProxy->pState, order + 1);
+		return (int) lua_tointeger(ctx->pProxy->pState, order + 1);
 	}
